test: extract tool policy white-box suite
